ov003: promote dScGameOver_c into a single C++ TU
Nine one-function shards folded into src/actors/dScGameOver_c.cpp, the
whole .text run 0x020b0580..0x020b1118. Byte-identical, whole range, from
one object.

  python tools/tubuild.py verify ov003/dScGameOver_c
    9/9 MATCH, objisolate clean, reloc-destinations clean -> TEXT-VERIFIED
    emission order: all 9 in ROM-ascending section order

  python tools/tubuild.py linkcheck ov003/dScGameOver_c   (pre-promotion)
    [4b/8] 9 LICENSED, 0 COLLIDES-GAP, 0 COLLIDES-SRC, 0 HOMELESS;
           emitted .text order ROM-ascending, as required
    [6/8]  TU .text range 2968 bytes IDENTICAL; module fidelity 106/106
           exact, 100.000000% of compared bytes; dsd check modules PASS
    [7/8]  dsd check symbols FAIL on the same 9 pre-existing errors the
           baseline control has -- 0 new, 0 resolved
    [8/8]  ROM built, sha256 d1506e90...c478e8, IDENTICAL to the stock
           control
    Result: SCRATCH-LINK-VERIFIED

THE DESTRUCTOR ORDER IS THE REPRODUCIBLE ONE. The cartridge puts D1 at
0x020b0580 BELOW D0 at 0x020b05bc, which is what inline-in-class emits, so
this TU promotes whole rather than holding the pair out. include/dScGameOver_c.h
moves `virtual ~dScGameOver_c();` to `virtual ~dScGameOver_c() {}` -- the
only admissible form that gives that order; out of line mwcc emits D2, D0,
D1 plus a D2 that exists nowhere in the ROM. No forcing scaffold is needed:
the inline destructor cannot be the key function, so InitResources is, and
this TU defines it -- which emits _ZTV13dScGameOver_c, whose slots 16 and 17
odr-use the pair and drag both out of line, D1 first. The two shards that
each carried an out-of-line definition of the same destructor are deleted,
not ported.

NINE compiler_only_output rows, all deadstrip-data, all with a configured
ROM home; externalized_output is empty. The count is what the chain predicts:
2 x (3 ancestors + self) + 1 vtable, with no Vector3 term -- this TU touches
none, as members or as locals. The eight typeinfo records and names are
dScGameOver_c's own (ov003:0x020b1758, 0x020b1784) plus dScene_c's, dBase_c's
and fBase_c's canonical arm9 copies. The vtable's extent is proven, not
inferred: 18 slots read word by word off the linked overlay, ending at
0x020b17e4 where the next word is not a code address in any loaded module.
`romdata_check --files src/actors/dScGameOver_c.cpp` reports 9 symbols,
6 VERIFIED, 3 PARTIAL, 0 DIFFERS; the vtable and both own RTTI records are
VERIFIED, and the 3 PARTIAL are the ancestor _ZTS strings.

Three hard conflicts the merge created that tubuild's detector does not
report, because it compares declaration spellings and not types:

  - data_0209d45c/data_0209d454 arrived as `unsigned char x[]` from one
    shard and `u8 x` from another. The scalar wins -- it is what the rest of
    the tree uses for these two BG-enable registers.
  - func_ov003_020b060c was defined `(char *)` against decl_common.h:2301's
    `(void*)`: `illegal function overloading`, reported against the
    definition line and never naming the header. The header wins; the
    `char *` view survives as a local alias so the raw-offset accesses keep
    their legacy spelling.
  - the legacy shadow types `VT`/`Obj` collided with decl_common.h:415's
    real ROM symbol `extern int VT[];`. A variable hides a struct tag in
    C++, so this surfaces as `undefined identifier 'VT'` on the struct that
    uses it, never on the header. Renamed GameOverGateVt/GameOverGate.

The one pragma blocker, `#pragma opt_strength_reduction off` from
func_ov003_020b060c.c, is CARRIED at file scope. It is file-global
last-wins, so there is no narrower place for it; the blast radius `create`
warns about was measured rather than assumed, and all 9 members MATCH with
it set.

No prose, port/ slice manifest or markdown reference named any deleted
shard, so steps 6 and 7 of the promotion plan are no-ops here.
check_tubuild_conflicts, layout_check and check_dead_references are clean,
and tiers_ratchet reports BACKSLID 0 with the four affected shards
correctly upgraded from path rows to member rows.
