+++ config/converted-baseline.json
@@ -705,10 +705,6 @@
     "src/_ZN13UpDownLiftBbh6RenderEv.cpp",
     "src/_ZN13UpDownLiftBbhD0Ev.cpp",
     "src/_ZN13UpDownLiftBbhD1Ev.cpp",
-    "src/_ZN13dScGameOver_c16CleanupResourcesEv.cpp",
-    "src/_ZN13dScGameOver_c16OnPendingDestroyEv.cpp",
-    "src/_ZN13dScGameOver_cD0Ev.cpp",
-    "src/_ZN13dScGameOver_cD1Ev.cpp",
     "src/_ZN13dScMgTeresa_c6RenderEv.cpp",
     "src/_ZN13dScMgTeresa_cD0Ev.cpp",
     "src/_ZN13dScMgTeresa_cD1Ev.cpp",
@@ -2146,6 +2142,10 @@
     "src/actors/dBgActor_c.cpp#_ZN10dBgActor_cC2Ev",
     "src/actors/dBgActor_c.cpp#_ZN10dBgActor_cD0Ev",
     "src/actors/dBgActor_c.cpp#_ZN10dBgActor_cD1Ev",
+    "src/actors/dScGameOver_c.cpp#_ZN13dScGameOver_c16CleanupResourcesEv",
+    "src/actors/dScGameOver_c.cpp#_ZN13dScGameOver_c16OnPendingDestroyEv",
+    "src/actors/dScGameOver_c.cpp#_ZN13dScGameOver_cD0Ev",
+    "src/actors/dScGameOver_c.cpp#_ZN13dScGameOver_cD1Ev",
     "src/actors/dScMgD3DBase_c.cpp#Sound_PlayBank1Panned",
     "src/actors/dScMgD3DBase_c.cpp#_ZN14dScMgD3DBase_c24OnHitByCannonBlastedCharEv",
     "src/actors/dScMgD3DBase_c.cpp#_ZN14dScMgD3DBase_cD0Ev",
