--- src/_ZN13dScGameOver_c16OnPendingDestroyEv.cpp
@@ -1,10 +1,0 @@
-//cpp
-// @symbol _ZN13dScGameOver_c16OnPendingDestroyEv
-/* recovered: real C++ method */
-/* dScGameOver_c::OnPendingDestroy() -- vtable slot 12. Nothing to unwind; the
- * slot exists only so dScene_c's teardown has something to call. */
-#include "dScGameOver_c.h"
-
-void dScGameOver_c::OnPendingDestroy()
-{
-}
